Fixed sourceSets dsl in example to compile for gradle versions > 4.10-rc-2.
Background:
As of 4.10-rc-2 there was a breaking change to the sourceSets kotlin dsl from project.java.sourceSets to just be sourceSets.

I found this because the minimum configuration example failed to compile for me on the latest version of gradle with the following error:

Unresolved reference. None of the following candidates is applicable because of receiver type mismatch:
public val Project.sourceSets: SourceSetContainer defined in org.gradle.kotlin.dsl
public val KotlinJvmProjectExtension.sourceSets: NamedDomainObjectContainer<KotlinSourceSet> defined in org.gradle.kotlin.dsl

This led me to https://docs.gradle.org/4.10-rc-2/release-notes.html#changes-to-the-gradle-kotlin-dsl and https://discuss.gradle.org/t/how-to-set-sourcesets-with-gradle-4-10-rc2/28166 which explains this breaking dsl change.