feat(): resolve JOOQ Constant by reflection to be able to work with any jooq version
Due to static access in code, the version defined in the plugin was ALWAYS used by the plugin instead of relying on the version used by "user project". By using reflection, we can bypass this and access to the value defined in Jooq Constants without any conflict.

Link to https://github.com/rohanprabhu/kotlin-dsl-gradle-jooq-plugin/issues/19