fix(cache): provide a stable confighash caclculation between runs
Due to the inconsistency of hashcode on objects between run, if we use remote cache or different gradle daemon for building (for example in CI with different steps), the `configHash` computation wasn't equals for the same configuration. Here, we rely on `JOOQ` stability on toString() for configuration (and also on stability of @lukaseder 😉).

Closes https://github.com/rohanprabhu/kotlin-dsl-gradle-jooq-plugin/issues/24